Remove unused TimeBudget entity helper
